Allow google to bypass checker

+++ public/header.js
@@ -13,7 +13,7 @@ if (!localStorage.getItem('name')) {
     fetch(`${base}users`)
       .then((res) => res.json())
       .then((body) => {
-        if (!localStorage.getItem('name')) return;
+        if (!localStorage.getItem('name') || localStorage.getItem('google')) return;
         if (!body.users.find(user => user[0] === localStorage.getItem('name') || user[1] === localStorage.getItem('id'))) {
           localStorage.clear();
           location.reload();
@@ -23,7 +23,7 @@ if (!localStorage.getItem('name')) {
   fetch(`${base}users`)
     .then((res) => res.json())
     .then((body) => {
-      if (!localStorage.getItem('name')) return;
+      if (!localStorage.getItem('name') || localStorage.getItem('google')) return;
       if (!body.users.find(user => user[0] === localStorage.getItem('name') || user[1] === localStorage.getItem('id'))) {
         localStorage.clear();
         location.reload();
